Collocate embedded artifact loaders
